Probably will remove
This is the original compare all levels file (taxaHFE vs SF vs L1-L7 vs etc). Now that we are moving to methods, we will end up just calling on the methods in the run file, depending on some flag parameters in the options file.

I think we also will have to think about how we repeat this process over random seeds. Should that just be a default choice (ie 5 random seeds)?