Changes to output paths
These may need to be tweaked further, but they work in my hands for now.

TODO: Yirui found a bug where, if there are NAs in the metadata file (not in the label column though! We would have gotten ride of rows with NAs in the metadata column), it will fail when it hits dietML, becase we have not given dietML any instructions on how to handle NAs. We need to drop those samples earlier probably.