Quick update to run_dietML_ranger()
Margaret stumbled on an error using dietML, which i think would transfer here. Previously, I let tune_bayes/ranger tune min_n (and I set bounds for mtry). It seems that in small sample size instances, min_n can be tuned to a number higher than your test size. This leads to problems when calculating the SHAP values...if your min_n is higher than the number of test samples you have, the SHAP plot will fail (youll get an error that ie You asked for 30 samples but you only have 26 samples). This is because min_n is larger than the test set. This seems to only affect the "test" plot for shap, and not train or full plots or anything else. However, we can put a gentle bound on min_n so this doesnt happen, but limited the upper end to the number of samples in the test dataset.

As for the lower bound, I played around with min_n = 1 and got an error, but not with min_n = 2 so i just went with 2.