Move where shap data is saved
put the saving of shap data inputs higher in the code. The trip point is the shap test data. We want the saving of the inputs to be before the saving of the test data, in case that triggers an error (before, in rare cases, the shap test data doesnt finish, and the loop will error out instead of write the inputs).