97 shap data save (#99)
* attempt to save RData

This code saves the .RData file even if the shap analysis doesnt finish successfully. Note i havent exhaustively tested this out, but it seems like it works in a few rare cases I could make. Also it feels like it should work since the shap_plot_env exists before the shap analysis is run...therefore the save should work, it would just save an empty enviornment if the shap analysis failed.

* Move where shap data is saved

put the saving of shap data inputs higher in the code. The trip point is the shap test data. We want the saving of the inputs to be before the saving of the test data, in case that triggers an error (before, in rare cases, the shap test data doesnt finish, and the loop will error out instead of write the inputs).