Update SHAP code to be safer
Fairly sure that the using dopar on the shap code will result in parallel explosion. ranger parallel * shap parallel. I think it literally ate through my computer (RIP). So we are
1) hard encoded parallel = F
2) dynamically encoding nsim to min out at 10 and max out at 200, and also it will error out if your input dataset is too high (ncol * nrow > 500,000)

Other things that happened is re-arranging the code so it writes things to the shap_plot_env as soon as theyre created. Also doing some rm and gc() at the end of the function to be safe