Based on evolution valley talking
We talked about how we should be not doing anything to test data except grabbing the features in the training dataset (which were processed with taxaHFE). I also added an if statement to error out if the column names do not match. Finally, i added some switch() logic so that even the abundance and prev filters are switched to zero when the test data is "run" (quotes becase you can see the other switches in this method section that basically barely run taxahfe...so its fast and produces taxa names which i can match between train and test. this is not necessary if we can somehow make the test data bail from compete_tree, but its also so fast that im not really worried about it).