Merge dev into main for a new version (v2.3.3) (#112)
* add parallel workers
* fix validators
* general flag validation and tests
* comment out skip too big shap run
Previously we prevented a shap analysis that was going to be way too long. Commenting it out because working with big data might be what taxahfe is used for and this is limiting...i still think this could be a warning or something.
* parallel workers matching github runner cores
* Based on evolution valley talking
We talked about how we should be not doing anything to test data except grabbing the features in the training dataset (which were processed with taxaHFE). I also added an if statement to error out if the column names do not match. Finally, i added some switch() logic so that even the abundance and prev filters are switched to zero when the test data is "run" (quotes becase you can see the other switches in this method section that basically barely run taxahfe...so its fast and produces taxa names which i can match between train and test. this is not necessary if we can somehow make the test data bail from compete_tree, but its also so fast that im not really worried about it).
* Add in args for method_taxa_hfe_ml
added args needed to specify writing_flattened_tree file (the "reasons" file), just for the training data.
* add write_flattened_tree func to taxahfe-ml method
inside loop where count==1, which means it only functions, if specified, on the training data.
* Converted shap error to warning
Previous commit commented this out to allow for larger analyses, but suggested we turn this into a warning. Which is what ive done. Future efforts may look into parallizing the SHAP analysis. I wish tree shap was possible too...
* Make shap parallelized
Used the info from the fastshap package (https://bgreenwell.github.io/fastshap/articles/fastshap.html) to parallelize the SHAP analysis. Used forking because it is a bit faster and limits overhead (but only works on POSIX machines...which i think is fine with our docker implementations). Updated shap_analysis() to take in parallel_workers arg instead of ncores.
* Update README.md
update taxahfe-ml options flags (eg now includes parallel_workers)

---------

Co-authored-by: Matt Kay <matthewckay@gmail.com>