remove all the correlation feature engineering in dietML
Previously we had all the old feature engineering in dietML (step_dummy, which is one-hot encoding, step_zv-remove zero variance-, and step corr). Since this is a feature engineering program, for the hierarchical features, i dont want any "extra" feature engineering happening downstream. It likely has no effect (except if the user adds highly correlated covariates, or zero var covariets), because the taxahfe cor threshold is already run in the taxahfe stage (and probs would not trigger the corr feature engineering in the dietML stage).

I also noticed a bug where, in the HP tuning model set, it was telling ranger to use total_cores instead of ncores. It can only use total_cores if it is NOT hyperparameter tuning. So that was a bug.