Add in novel() step for recipe
This step is vital for datasets which may contain a factor with one level being VERY sparse. If that level never ends up in the training subset, then it is not dummy encoded properly. Then the test set basically sees an NA for that sample in one of the dummy encoded levels that the training data did see. An NA in the test data then causes failure. This is admittedly a kind of weird bug - you really only have 1-2 instances of a level within a factor? Probs not a great factor...