Simplify taxahfe_ml and summarized levels methods
Get rid of compete tree run twice (on train and test data for taxahfe_ml). This is because prepare_flattened_df can actually take in hTree too. We dont compete the test data because all features in the test data are used in the ML (well, actually only the features that appear in the train data. The models dont know what to do with features that they are not trained on, so they ignore those.).

Likewise, method_levels doesnt need to compete the tree either, so we save a lot of time there too by just passing hTree to the flatten function.