More changes to make test_options happy
I had a min_warning set for info_gain_n, but that value was also the default. I dont think that worked, because the warning never showed up and since test_options was looking for that warning, it failed. I just added more info in the help message for the user, i dont think the min_warning (but also default) is necessary and obviously its problamatic.