updates to docker install and pkg versions
update to R v4.5.2 and all packages. The way I did this was to download an rstudio image of R4.5.2 and, in the taxahfe folder, remove all renv folders and the renv.lock. Then i reinstalled everything and made a fresh lock file. Probs not the most ideal way to do that.